BernsteinBasis

Documentation for BernsteinBasis.

BernsteinBasis.BernsteinDerivativeMatrix_3D_r

BernsteinBasis.BernsteinDerivativeMatrix_3D_s

BernsteinBasis.BernsteinDerivativeMatrix_3D_t

BernsteinBasis.BernsteinLift

BernsteinBasis.ElevationMatrix

BernsteinBasis.MultithreadedBernsteinLift

BernsteinBasis.bernstein_2d_scalar_multiindex_lookup

BernsteinBasis.bernstein_basis

BernsteinBasis.bernstein_basis_from_barycentric

BernsteinBasis.cartesian_to_barycentric

BernsteinBasis.elevation_multiply!

BernsteinBasis.face_lift_multiply!

BernsteinBasis.get_bernstein_lift

BernsteinBasis.get_coeff

BernsteinBasis.ij_to_linear

BernsteinBasis.ijk_to_linear

BernsteinBasis.l_j

BernsteinBasis.reduction_multiply!

BernsteinBasis.threaded_mul!

LinearAlgebra.mul!

BernsteinBasis.BernsteinDerivativeMatrix_3D_r — Type

BernsteinDerivativeMatrix_3D_r <: AbstractMatrix{Float64}

Derivative matrix with respect to the first Cartesian coordinate $r$ in the 3D Bernstein basis.

Fields

N::Int: Order of Bernstein polynomials. Supports up to $N = 20$.

BernsteinBasis.BernsteinDerivativeMatrix_3D_s — Type

BernsteinDerivativeMatrix_3D_s <: AbstractMatrix{Float64}

Derivative matrix with respect to the second Cartesian coordinate $s$ in the 3D Bernstein basis.

Fields

N::Int: Order of Bernstein polynomials. Supports up to $N = 20$.

BernsteinBasis.BernsteinDerivativeMatrix_3D_t — Type

BernsteinDerivativeMatrix_3D_t <: AbstractMatrix{Float64}

Derivative matrix with respect to the third Cartesian coordinate $t$ in the 3D Bernstein basis.

Fields

N::Int: Order of Bernstein polynomials. Supports up to $N = 20$.

BernsteinBasis.BernsteinLift — Type

BernsteinLift

Lift matrix for a single face on a standard tetrahedron. Supports mul! operations, but should optimally be used for matrix-vector multiplications rather than matrix-matrix multiplications.

Chan, Jesse and Tim Warburton (2017) GPU-accelerated Bernstein-Bezier discontinuous Galerkin methods for wave problems DOI: 10.48550/arXiv.1512.06025

BernsteinBasis.ElevationMatrix — Type

ElevationMatrix{N} <: AbstractMatrix{Float64}

Two-dimensional degree elevation operator for Bernstein polynomials. Expresses polynomials of degree $N - 1$ as polynomials of degree $N$.

Kirby, Robert C. (2016) Fast inversion of the simplicial Bernstein mass matrix DOI: 10.1007/s00211-016-0795-0

BernsteinBasis.MultithreadedBernsteinLift — Type

MultithreadedBernsteinLift

A multithreaded-friendly lift matrix for a single face on a standard tetrahedron. Supports mul! operations, but should optimally be used for matrix-vector multiplications rather than matrix-matrix multiplications.

Chan, Jesse and Tim Warburton (2017) GPU-accelerated Bernstein-Bezier discontinuous Galerkin methods for wave problems DOI: 10.48550/arXiv.1512.06025

BernsteinBasis.bernstein_2d_scalar_multiindex_lookup — Method

bernstein_2d_scalar_multiindex_lookup(N)

Returns a vector that maps scalar indices to multi-indices of the $N$-th degree Bernstein polynomials.

BernsteinBasis.bernstein_basis — Method

bernstein_basis(::Line, N, r)
bernstein_basis(elem::Tri, N, r, s)
bernstein_basis(elem::Tet, N, r, s, t)

Wrapper for bernstein_basis_from_barycentric that takes in $rst$-space coordinates, each in $[-1, 1]$, instead of barycentric coordinates.

Returns the $N_p x N_p$ generalized Vandermonde matrix $\mathcal{V}$, where $\mathcal{V}_{mn}$ = the $n$-th Bernstein basis evaluated at the $m$-th point, followed by its derivative matrices.

Use ::Line for 1D, ::Tri for 2D, and ::Tet for 3D.

We order the Bernstein basis according to exponents in a "reverse-dictionary" order of the first DIM-1 coordinates. Does not work for $N > 20$ because of factorial() limitations.

Arguments

N::Int: Bernstein basis degree

r,s,t::AbstractArray{T,1}: $N_p$-sized vectors of distinct 2D barycentric coordinates to be

used as interpolatory points

BernsteinBasis.bernstein_basis_from_barycentric — Method

bernstein_basis_from_barycentric(::Line, N, r, s)
bernstein_basis_from_barycentric(::Tri, N, r, s, t)
bernstein_basis_from_barycentric(::Tet, N, r, s, t, u)

Returns the $N_p x N_p$ generalized Vandermonde matrix $\mathcal{V}$, where $\mathcal{V}_{mn}$ = the $n$-th Bernstein basis evaluated at the $m$-th point, followed by its derivative matrices.

Use ::Line for 1D, ::Tri for 2D, and ::Tet for 3D.

We order the Bernstein basis according to exponents in a "reverse-dictionary" order of the first DIM-1 coordinates. Does not work for $N > 20$ because of factorial() limitations.

Arguments

N::Int: Bernstein basis degree

r,s,t::AbstractArray{T,1}: $N_p$-sized vectors of distinct 2D barycentric coordinates to be

used as interpolatory points

BernsteinBasis.cartesian_to_barycentric — Method

cartesian_to_barycentric(elem::Union{Line, Tri, Tet}, coords)
cartesian_to_barycentric(elem::Union{Line, Tri, Tet}, coords...)

Converts a DIMS by $N_p$ matrix of cartesian coordinates into a matrix of barycentric coordinates. Alternatively, also takes in DIMS number of vectors of coordinates.

BernsteinBasis.elevation_multiply! — Method

elevation_multiply!(out, N, x, offset)

Multiplies x by the 2D Bernstein elevation matrix that maps from degree $N$ polynomials to $N + 1$.

tri_offsets(N) should be passed as the precomputed offset value.

BernsteinBasis.face_lift_multiply! — Method

fast_lift_multiply!(out, N, L0, x, offset, l_j, E)

Multiplies x by the "nice" lift matrix face ($rs$-plane) of the lift matrix.

Arguments

x::AbstractVector: Input vector

Chan, Jesse and Tim Warburton (2017) GPU-accelerated Bernstein-Bezier discontinuous Galerkin methods for wave problems DOI: 10.48550/arXiv.1512.06025

BernsteinBasis.get_bernstein_lift — Method

get_bernstein_lift(N)

Returns an $N$-degree 3D Bernstein lift matrix as a Matrix type, in contrast to the more optimized BernsteinLift(N).

This function is derived from StartUpDG's Tet() struct.

BernsteinBasis.get_coeff — Method

get_coeff(i1, j1, k1, l1, i2, j2, k2, l2)

Returns the value of the (i1, j1, k1, l1), (i2, j2, k2, l2)-th entry of the 3D Bernstein derivative matrix with respect to $i$ (first barycentric coordinate).

BernsteinBasis.ij_to_linear — Method

ij_to_linear(i,j,offset)

Returns the scalar index of the $(i,j,k)$-th 2D Bernstein basis.

BernsteinBasis.ijk_to_linear — Method

ijk_to_linear(i,j,offset)

Returns the scalar index of the $(i,j,k,l)$-th 3D Bernstein basis.

BernsteinBasis.l_j — Method

l_j(N)

Returns a tuple containing $l_1, ..., l_N, ..., l_{20}$ as specified by the lift matrix multiplication algorithm in

Chan, Jesse and Tim Warburton (2017) GPU-accelerated Bernstein-Bezier discontinuous Galerkin methods for wave problems DOI: 10.48550/arXiv.1512.06025

BernsteinBasis.reduction_multiply! — Method

reduction_multiply!(out, N, x, offset)

Multiplies x by the 2D Bernstein reduction matrix that maps from degree $N$ polynomials to $N - 1$.

tri_offsets(N) should be passed as the precomputed offset value.

BernsteinBasis.threaded_mul! — Method

threaded_mul!(out::AbstractVector, L::BernsteinLift, x::AbstractVector)

Multiplies x by the Bernstein lift matrix L, as defined in documentation.

Arguments

out::AbstractVector: Length $rac{(N+1)(N+2)(N+3)}{6}$ vector

L::BernsteinLift: Order $N$ Bernstein lift matrix

x::AbstractVector: Length $2 \cdot rac{(N+1)(N+2)}{2}$ vector

thread: Thread id

LinearAlgebra.mul! — Method

LinearAlgebra.mul!(out::AbstractVector, L::BernsteinLift, x::AbstractVector)

Multiplies x by the Bernstein lift matrix L, as defined in documentation.

Arguments

out::AbstractVector: Length $rac{(N+1)(N+2)(N+3)}{6}$ vector

L::BernsteinLift: Order $N$ Bernstein lift matrix

x::AbstractVector: Length $2 \cdot rac{(N+1)(N+2)}{2}$ vector