BernsteinBasis

Documentation for BernsteinBasis.

BernsteinBasis.BernsteinDerivativeMatrix_3D_r

BernsteinBasis.BernsteinDerivativeMatrix_3D_s

BernsteinBasis.BernsteinDerivativeMatrix_3D_t

BernsteinBasis.BernsteinLift

BernsteinBasis.ElevationMatrix

BernsteinBasis.bernstein_2d_scalar_multiindex_lookup

BernsteinBasis.fast_lift_multiply!

BernsteinBasis.get_coeff

BernsteinBasis.ij_to_linear

BernsteinBasis.ijk_to_linear

BernsteinBasis.reduction_multiply!

BernsteinBasis.BernsteinDerivativeMatrix_3D_r — Type

BernsteinDerivativeMatrix_3D_r <: AbstractMatrix{Float64}

Derivative matrix with respect to the first Cartesian coordinate $r$ in the 3D Bernstein basis.

Fields

N::Int: Order of Bernstein polynomials. Supports up to $N = 20$.

BernsteinBasis.BernsteinDerivativeMatrix_3D_s — Type

BernsteinDerivativeMatrix_3D_s <: AbstractMatrix{Float64}

Derivative matrix with respect to the second Cartesian coordinate $s$ in the 3D Bernstein basis.

Fields

N::Int: Order of Bernstein polynomials. Supports up to $N = 20$.

BernsteinBasis.BernsteinDerivativeMatrix_3D_t — Type

BernsteinDerivativeMatrix_3D_t <: AbstractMatrix{Float64}

Derivative matrix with respect to the third Cartesian coordinate $t$ in the 3D Bernstein basis.

Fields

N::Int: Order of Bernstein polynomials. Supports up to $N = 20$.

BernsteinBasis.BernsteinLift — Type

BernsteinLift

Lift matrix for a single face on a standard tetrahedron.

Chan, Jesse and Tim Warburton (2017) GPU-accelerated Bernstein-Bezier discontinuous Galerkin methods for wave problems DOI: 10.48550/arXiv.1512.06025

BernsteinBasis.ElevationMatrix — Type

ElevationMatrix{N} <: AbstractMatrix{Float64}

Two-dimensional degree elevation operator for Bernstein polynomials. Expresses polynomials of degree $N - 1$ as polynomials of degree $N$.

Kirby, Robert C. (2016) Fast inversion of the simplicial Bernstein mass matrix DOI: 10.1007/s00211-016-0795-0

BernsteinBasis.bernstein_2d_scalar_multiindex_lookup — Method

bernstein_2d_scalar_multiindex_lookup(N)

Returns a vector that maps scalar indices to multi-indices of the $N$-th degree Bernstein polynomials.

BernsteinBasis.fast_lift_multiply! — Method

fast_lift_multiply!(out, N, L0, x, offset, l_j, E)

Multiplies x by the "nice" lift matrix face ($rs$-plane).

Arguments

$L0::SparseMatrixCSC{Float64, Int64}$: Precomputed matrix as defined in (Chan 2017) $x::AbstractVector$: Input vector $tri_offset_table::NTuple{20, NTuple{21, Int}}$: Precomputed vector of offset tuples $l_j_table::NTuple{20, Float64}$: Precomputed coefficients given by l_j(N) $E::Vector{Float64}$ Dummy vector (with same dimensions as x) to reduce redundant memory allocation.

Chan, Jesse and Tim Warburton (2017) GPU-accelerated Bernstein-Bezier discontinuous Galerkin methods for wave problems DOI: 10.48550/arXiv.1512.06025

BernsteinBasis.get_coeff — Method

get_coeff(i1, j1, k1, l1, i2, j2, k2, l2)

Returns the value of the (i1, j1, k1, l1), (i2, j2, k2, l2)-th entry of the 3D Bernstein derivative matrix with respect to $i$ (first barycentric coordinate).

BernsteinBasis.ij_to_linear — Method

ij_to_linear(i,j,offset)

Returns the scalar index of the $(i,j,k)$-th 2D Bernstein basis.

BernsteinBasis.ijk_to_linear — Method

ijk_to_linear(i,j,offset)

Returns the scalar index of the $(i,j,k,l)$-th 3D Bernstein basis.

BernsteinBasis.reduction_multiply! — Method

reduction_multiply!(out, N, x, offset)

Multiplies x by the Bernstein reduction matrix that maps from degree $N$ polynomials to $N - 1$.

tri_offsets(N) should be passed as the precomputed offset value.