BernsteinBasis

Documentation for BernsteinBasis.

BernsteinBasis.BernsteinDerivativeMatrix_3D_r

BernsteinBasis.BernsteinDerivativeMatrix_3D_s

BernsteinBasis.BernsteinDerivativeMatrix_3D_t

BernsteinBasis.get_coeff

BernsteinBasis.BernsteinDerivativeMatrix_3D_r — Type

BernsteinDerivativeMatrix_3D_r{N} <: AbstractMatrix{Float64}

Derivative matrix with respect to the first Cartesian coordinate r in the 3D Bernstein basis.

# Fields
- `N::Int`: Order of Bernstein polynomials. Supports up to N = 20.

BernsteinBasis.BernsteinDerivativeMatrix_3D_s — Type

BernsteinDerivativeMatrix_3D_s{N} <: AbstractMatrix{Float64}

Derivative matrix with respect to the second Cartesian coordinate s in the 3D Bernstein basis.

# Fields
- `N::Int`: Order of Bernstein polynomials. Supports up to N = 20.

BernsteinBasis.BernsteinDerivativeMatrix_3D_t — Type

BernsteinDerivativeMatrix_3D_t{N} <: AbstractMatrix{Float64}

Derivative matrix with respect to the third Cartesian coordinate t in the 3D Bernstein basis.

# Fields
- `N::Int`: Order of Bernstein polynomials. Supports up to N = 20.

BernsteinBasis.get_coeff — Method

get_coeff(i1, j1, k1, l1, i2, j2, k2, l2)

Returns the value of the (i1, j1, k1, l1), (i2, j2, k2, l2)-th entry of the 3D Bernstein derivative matrix with respect to i (first barycentric coordinate).